Share chunk transformer.
